Move process classes to separate files
